src/torchserve_sdxl/fastapi/requirements.txt
fastapi[all] 
boto3
numpy
requests
pillow